tfix nodes.test.ts ESM load crash
