+++ packages/server/src/services/nodes/index.ts
@@ -2,13 +2,16 @@ import { cloneDeep, omit } from 'lodash'
 import { StatusCodes } from 'http-status-codes'
 import { getRunningExpressApp } from '../../utils/getRunningExpressApp'
 import { INodeData, MODE } from '../../Interface'
-import { INode, INodeOptionsValue, INodeParams, ClientType } from 'flowise-components'
+import { ClientType, INodeOptionsValue } from 'flowise-components'
 import { databaseEntities } from '../../utils'
 import logger from '../../utils/logger'
 import { InternalFlowiseError } from '../../errors/internalFlowiseError'
 import { getErrorMessage } from '../../errors/utils'
 import { OMIT_QUEUE_JOB_DATA } from '../../utils/constants'
 import { executeCustomNodeFunction } from '../../utils/executeCustomNodeFunction'
+import { filterNodeByClient } from './filterNodeByClient'
+
+export { filterNodeByClient }
 
 // Get all component nodes
 const getAllNodes = async (client?: ClientType) => {
@@ -150,29 +153,6 @@ const executeCustomFunction = async (requestBody: any, workspaceId?: string, org
     } else {
         return await executeCustomNodeFunction(executeData)
     }
-}
-
-// Filter node inputs by client. Params/options with a `client` array that excludes the requesting client are removed. No-ops when client is omitted.
-export const filterNodeByClient = (node: INode, client?: ClientType): INode => {
-    if (!client || !node.inputs) return node
-
-    const filterParam = (param: INodeParams): INodeParams => {
-        const filtered: INodeParams = { ...param }
-        if (filtered.options) {
-            filtered.options = filtered.options.filter((opt: INodeOptionsValue) => !opt.client || opt.client.includes(client))
-        }
-        if (filtered.tabs) {
-            filtered.tabs = filtered.tabs.filter((t) => !t.client || t.client.includes(client)).map(filterParam)
-        }
-        if (filtered.array) {
-            filtered.array = filtered.array.filter((a) => !a.client || a.client.includes(client)).map(filterParam)
-        }
-        return filtered
-    }
-
-    const filteredInputs = (node.inputs as INodeParams[]).filter((param) => !param.client || param.client.includes(client)).map(filterParam)
-
-    return { ...node, inputs: filteredInputs }
 }
 
 export default {
